+++ R/load_forecasts_zoltar.R
@@ -17,7 +17,9 @@
 #' Default to all valid targets in Zoltar.
 #' @param as_of character for date time to load forecasts submitted as of this time. 
 #' It could use the format of one of the three examples: 
-#' "2021-01-01", "2020-01-01 01:01:01" and "2020-01-01 01:01:01 EDT".
+#' "2021-01-01", "2020-01-01 01:01:01" and "2020-01-01 01:01:01 UTC".
+#' If you would like to set a timezone, it has to be UTC now. 
+#' If not, helper function will append the default timezone to your input based on hub parameter. 
 #' Default to NULL to load the latest version.
 #' @param hub character vector, where the first element indicates the hub
 #' from which to load forecasts. Possible options are "US" and "ECDC"
@@ -39,7 +41,7 @@ load_forecasts_zoltar <- function (
   verbose = TRUE) {
   
   # set up Zoltar connection
-  zoltar_connection <- setup_zoltar_connection()
+  zoltar_connection <- setup_zoltar_connection(staging = TRUE)
   
   if (!is.null(forecast_dates)){
     # construct Zoltar project url
@@ -67,7 +69,7 @@ load_forecasts_zoltar <- function (
                                       targets = targets,
                                       types = types,
                                       verbose = verbose,
-                                      as_of = date_to_datetime(as_of))
+                                      as_of = date_to_datetime(as_of, hub))
   if (nrow(forecasts) == 0){
     warning("Warning in do_zoltar_query: Forecasts are not available.\n Please check your parameters.")
     # convert value column to double and select columns