+++ R/load_latest_forecasts_zoltar.R
@@ -18,7 +18,9 @@
 #' Default to all valid targets in Zoltar.
 #' @param as_of character for date time to load forecasts submitted as of this time. 
 #' It could use the format of one of the three examples: 
-#' "2021-01-01", "2020-01-01 01:01:01" and "2020-01-01 01:01:01 EDT".
+#' "2021-01-01", "2020-01-01 01:01:01" and "2020-01-01 01:01:01 UTC".
+#' If you would like to set a timezone, it has to be UTC now. 
+#' If not, helper function will append the default timezone to your input based on hub parameter.
 #' Default to NULL to load the latest version.
 #' @param verbose a boolean for printing messages on zoltar job status. Default to TRUE.
 #' @param hub character vector, where the first element indicates the hub