+++ NEWS.md
@@ -6,7 +6,7 @@
 - Add a column that appends state abbreviation to county names in US hub locations data object.
 - `load_truth()` now supports multiple target variables and has a new set of default values for `target_variable` and `truth_source` based on `hub` parameter.
 - `load_forecasts_repo()`, `load_latest_forecasts_repo()` and `load_forecast_files_repo()` drop rows with NULLs in `value` column in forecast files.
-
+- Add `date_to_datetime()` that converts a date to a date time with timezone offset for `as_of` parameter. This function is used when the user is loading forecasts from zoltar only.
 
 ## covidHubUtils 0.1.5